pyproject.toml
[project]
name = "agent-first-service"
version = "0.0.1"
description = "Agent-first services demo (Projects/People/Communications + Orchestrator)."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "pydantic>=2.9",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "langgraph>=0.2,<0.4",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.2,<0.4",
    "sse-starlette>=2.1",
    "mcp>=1.6,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent_protocol*", "services*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "live: exercises a live external service (e.g. Azure OpenAI). Costs real tokens; keep in the default suite but make filterable for CI.",
]

[tool.ruff]
line-length = 100
